+++ src/ts/components/registry.ts
@@ -31,7 +31,6 @@ export const AUTO_INIT_COMPONENTS = {
   Sidenav: { component: Components.Sidenav, selector: '.sidenav' },
   NavigationRail: { component: Components.NavigationRail, selector: '.navigation-rail' },
   Tabs: { component: Components.Tabs, selector: '.tabs' },
-  TapTarget: { component: Components.TapTarget, selector: '.tap-target' },
   Timepicker: { component: Components.Timepicker, selector: '.timepicker' },
   Tooltip: { component: Components.Tooltip, selector: '.tooltipped' },
   FloatingActionButton: {